modified ClientManager for new entityManager structure.
Now we listen for events of the entity manager, and broadcast to the clients from the client manager, instead of from inside the entity manager. This also means that clients that do not have an entity loaded will still get entity load/unload events.

Additionally, the class was upgraded to ES6, and the socketWrapper class was made superfluous.